feat: 스티커 딤드 처리 색상 변경
