+++ DonDog-iOS/DonDog-iOS/Presentation/Features/Camera/Views/CameraView.swift
@@ -16,7 +16,7 @@ struct CameraView: UIViewControllerRepresentable {
     func makeUIViewController(context: Context) -> UIViewController {
         let customCameraVC = CustomCameraViewController()
         customCameraVC.delegate = context.coordinator
-        
+        customCameraVC.viewModel = viewModel
         customCameraVC.isStickerCamera = viewModel.isStickerCamera
         if let keyword = viewModel.stickerKeyword {
             customCameraVC.stickerKeyword = keyword
@@ -54,9 +54,7 @@ struct CameraView: UIViewControllerRepresentable {
         }
         
         func didCompleteBothPhotos() {
-            DispatchQueue.main.async {
-                self.parent.viewModel.showCaptionView = true
-            }
+            
         }
         
         func didCancel() {
